requirements.txt
playwright==1.58.0
openai==2.21.0
python-dotenv==1.2.1
